* Remove special host class
The mapping between the various platform strings was very finicky and
unreiable. This commit replaces that system with a simple map of
beaker-hostgenerator strings to vanagon targets.

* Disallow using hostnames for agent and master

This functionality might be added back later on, but it relied in a
weird way on the Host class that this commit removes, so it should be
reimplemented from scratch in a better way later.